# public/openapi.yaml
openapi: 3.0.3
info:
  title: Fine-Tune Labs API
  version: 1.0.0
  description: |

    # Fine-Tune Labs Public API

    Production-grade API for LLMOps, model training, and monitoring.

    ## Features

    - 🎯 **Widget Apps** - Deploy production monitoring widgets
    - 📊 **Analytics & Metrics** - Real-time model performance tracking
    - 🚀 **Inference Deployment** - Deploy and manage model endpoints
    - 🤖 **Model Management** - CRUD operations for LLM models
    - 📈 **Training Jobs** - Monitor fine-tuning progress

    ## Authentication

    All endpoints require Bearer token authentication:

    ```
    Authorization: Bearer YOUR_API_TOKEN
    ```

    Get your API token from the Fine-Tune Labs dashboard.

    ## Base URL

    - **Development**: `http://localhost:3000`
    - **Production**: `https://finetunelab.ai`

    ## Rate Limits

    - 1000 requests per minute per API key
    - 10,000 requests per day per API key

    ## Support

    - Documentation: https://finetunelab.ai/docs
    - Email: support@finetunelab.ai
  contact:
    name: Fine-Tune Labs API Support
    email: support@finetunelab.ai
    url: https://finetunelab.ai/support
  license:
    name: Proprietary
    url: https://finetunelab.ai/terms
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://finetunelab.ai
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        code:
          type: string
          description: Error code
      required:
        - error
...
security:
  - bearerAuth: []
...
paths:
...
  /api/training:
    get:
      summary: List training configurations
      description: |
        Retrieve all training configurations for the authenticated user.

        Training configurations define the hyperparameters, model architecture,
        and training settings for fine-tuning jobs.

        **Use Cases:**
        - List all saved training configurations
        - Select a config to start a training job
        - Review previously used training settings
      tags:
        - Training
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Configurations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  configs:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: cfg_abc123
                        name:
                          type: string
                          example: Llama 2 7B Fine-tune
                        description:
                          type: string
                          example: Customer support chatbot training config
                        template_type:
                          type: string
                          enum:
                            - lora
                            - full_finetune
                            - qlora
                            - dpo
                            - rlhf
                          example: lora
                        config_json:
                          type: object
                          description: Training hyperparameters and settings
                        is_validated:
                          type: boolean
                          example: true
                        created_at:
                          type: string
                          format: date-time
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create training configuration
      description: |
        Create a new training configuration with hyperparameters and settings.

        Training configurations are validated before being saved. Invalid configs
        will still be created but flagged with validation errors.

        **Use Cases:**
        - Save reusable training configurations
        - Create configs for different model types or use cases
        - Template configs for team members

        **Supported Template Types:**
        - `lora` - LoRA (Low-Rank Adaptation) fine-tuning
        - `qlora` - QLoRA (Quantized LoRA) for memory efficiency
        - `full_finetune` - Full parameter fine-tuning
        - `dpo` - Direct Preference Optimization
        - `rlhf` - Reinforcement Learning from Human Feedback
      tags:
        - Training
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - template_type
                - config_json
              properties:
                name:
                  type: string
                  description: Configuration name
                  example: Llama 2 7B LoRA
                description:
                  type: string
                  description: Optional description
                  example: LoRA fine-tuning for customer support chatbot
                template_type:
                  type: string
                  enum:
                    - lora
                    - full_finetune
                    - qlora
                    - dpo
                    - rlhf
                  example: lora
                config_json:
                  type: object
                  description: Training hyperparameters
                  properties:
                    model_name:
                      type: string
                      example: meta-llama/Llama-2-7b-hf
                    training:
                      type: object
                      properties:
                        learning_rate:
                          type: number
                          example: 0.00002
                        epochs:
                          type: integer
                          example: 3
                        batch_size:
                          type: integer
                          example: 4
                        warmup_ratio:
                          type: number
                          example: 0.1
                    lora:
                      type: object
                      properties:
                        r:
                          type: integer
                          example: 16
                        alpha:
                          type: integer
                          example: 32
                        dropout:
                          type: number
                          example: 0.05
      responses:
        '201':
          description: Configuration created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      is_validated:
                        type: boolean
                      validation_errors:
                        type: array
                        items:
                          type: string
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...